+++ apps/contrib/mixins.py
@@ -2,6 +2,7 @@ from django import forms
 from django.contrib.auth.mixins import LoginRequiredMixin
 from django.contrib.auth.mixins import UserPassesTestMixin
 from django.utils.translation import gettext_lazy as _
+from guest_user.functions import is_guest_user
 
 
 class StaffRequiredMixin(LoginRequiredMixin, UserPassesTestMixin):
@@ -10,6 +11,7 @@ class StaffRequiredMixin(LoginRequiredMixin, UserPassesTestMixin):
     def test_func(self):
         return self.request.user.is_active and self.request.user.is_staff
 
+
 RIGHT_OF_USE_LABEL = _(
     "I hereby confirm that the copyrights for this "
     "photo are with me or that I have received "
@@ -17,6 +19,15 @@ RIGHT_OF_USE_LABEL = _(
     "that the privacy rights of depicted third persons "
     "are not violated. "
 )
+
+
+class GuestCreatorContactFieldMixin:
+    """Keep creator contact email empty for guest users on create forms."""
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        if self.user and is_guest_user(self.user):
+            self.fields["creator_email"].initial = ""
 
 
 class ImageRightOfUseMixin(forms.ModelForm):