+++ apps/ideas/forms.py
@@ -4,6 +4,7 @@ from adhocracy4.categories.forms import CategorizableFieldMixin
 from adhocracy4.forms.fields import CreatorContactFieldMixin
 from adhocracy4.labels.mixins import LabelsAddableFieldMixin
 from apps.contrib.image_upload_help import IMAGE_UPLOAD_IDEA_HELP_TEXT
+from apps.contrib.mixins import GuestCreatorContactFieldMixin
 from apps.contrib.mixins import ImageRightOfUseMixin
 from apps.organisations.mixins import OrganisationTermsOfUseMixin
 
@@ -11,6 +12,7 @@ from . import models
 
 
 class IdeaForm(
+    GuestCreatorContactFieldMixin,
     CategorizableFieldMixin,
     LabelsAddableFieldMixin,
     ImageRightOfUseMixin,