+++ apps/mapideas/forms.py
@@ -6,6 +6,7 @@ from adhocracy4.forms.fields import CreatorContactFieldMixin
 from adhocracy4.labels.mixins import LabelsAddableFieldMixin
 from adhocracy4.maps import widgets as maps_widgets
 from apps.contrib.image_upload_help import IMAGE_UPLOAD_IDEA_HELP_TEXT
+from apps.contrib.mixins import GuestCreatorContactFieldMixin
 from apps.contrib.mixins import ImageRightOfUseMixin
 from apps.organisations.mixins import OrganisationTermsOfUseMixin
 
@@ -13,6 +14,7 @@ from . import models
 
 
 class MapIdeaForm(
+    GuestCreatorContactFieldMixin,
     CategorizableFieldMixin,
     LabelsAddableFieldMixin,
     ImageRightOfUseMixin,