+++ apps/users/constants.py
@@ -1,0 +1,1 @@
+GUEST_SWITCH_QUERY_PARAM = "guest_switch"